Move main() function out of mscore directory
