+++ CMakeLists.txt
@@ -739,11 +739,73 @@ if (NOT MSVC)
 endif (NOT MSVC)
 
 ##
+## Includes
+##
+include_directories(
+   ${PROJECT_SOURCE_DIR}
+   ${PROJECT_BINARY_DIR}
+   ${ALSA_INCDIR}
+   ${JACK_INCDIR}
+   ${PORTAUDIO_INCDIR}
+   ${OGG_INCDIR}
+   ${VORBIS_INCDIR}
+   ${SNDFILE_INCDIR}
+   ${LAME_INCLUDE_DIR}
+)
+
+if (MSVC)
+   include_directories(${PROJECT_SOURCE_DIR}/dependencies)
+   include_directories(${PROJECT_SOURCE_DIR}/dependencies/include)
+   include_directories(${PROJECT_SOURCE_DIR}/dependencies/include/vorbis)
+   include_directories(${SNDFILE_INCLUDE_DIR})
+   include_directories(${VORBIS_INCLUDE_DIR})
+   include_directories(${PORTAUDIO_INCLUDE_DIR})
+   include_directories(${OGG_INCLUDE_DIR})
+endif (MSVC)
+
+if (USE_SYSTEM_FREETYPE)
+      include_directories(${FREETYPE_INCLUDE_DIRS})
+else (USE_SYSTEM_FREETYPE)
+      include_directories(${PROJECT_SOURCE_DIR}/thirdparty/freetype/include)
+endif (USE_SYSTEM_FREETYPE)
+
+if (OMR)
+ if (NOT USE_SYSTEM_POPPLER)
+   include_directories(${PROJECT_SOURCE_DIR}/thirdparty/poppler)
+   include_directories(${PROJECT_SOURCE_DIR}/thirdparty/poppler/poppler)
+ else (NOT USE_SYSTEM_POPPLER)
+   include_directories(/usr/include/poppler)
+   include_directories(/usr/include/poppler/qt5)
+ endif (NOT USE_SYSTEM_POPPLER)
+endif (OMR)
+
+add_subdirectory(mscore)
+add_subdirectory(main)
+
+if (BUILD_CRASH_REPORTER)
+      set (CRASHREPORTER_EXECUTABLE_NAME "${MSCORE_EXECUTABLE_NAME}-crash-reporter")
+      add_subdirectory(crashreporter)
+      add_dependencies(mscore mscore-crash-reporter)
+      set (CRASHREPORTER_EXECUTABLE "${CRASHREPORTER_EXECUTABLE_NAME}${CMAKE_EXECUTABLE_SUFFIX}")
+      set (MSCORE_EXECUTABLE "${MSCORE_EXECUTABLE_NAME}${CMAKE_EXECUTABLE_SUFFIX}")
+endif (BUILD_CRASH_REPORTER)
+
+##
+## produce config.h file
+##
+configure_file (
+      ${PROJECT_SOURCE_DIR}/build/config.h.in
+      ${PROJECT_BINARY_DIR}/config.h
+      )
+
+##
 ## Add subdirs
 ##
+add_subdirectory(effects)
+add_subdirectory(synthesizer)
 subdirs(
-      awl bww2mxml share midi audiofile fluid libmscore synthesizer
-      effects thirdparty/rtf2html thirdparty/beatroot
+      awl bww2mxml share midi audiofile fluid libmscore
+      thirdparty/rtf2html thirdparty/beatroot
       thirdparty/qzip thirdparty/kQOAuth
       )
 
@@ -792,65 +854,6 @@ endif (OSC)
 if (NOT USE_SYSTEM_FREETYPE)
       subdirs (thirdparty/freetype)
 endif (NOT USE_SYSTEM_FREETYPE)
-
-##
-## Includes
-##
-include_directories(
-   ${PROJECT_SOURCE_DIR}
-   ${PROJECT_BINARY_DIR}
-   ${ALSA_INCDIR}
-   ${JACK_INCDIR}
-   ${PORTAUDIO_INCDIR}
-   ${OGG_INCDIR}
-   ${VORBIS_INCDIR}
-   ${SNDFILE_INCDIR}
-   ${LAME_INCLUDE_DIR}
-)
-
-if (MSVC)
-   include_directories(${PROJECT_SOURCE_DIR}/dependencies)
-   include_directories(${PROJECT_SOURCE_DIR}/dependencies/include)
-   include_directories(${PROJECT_SOURCE_DIR}/dependencies/include/vorbis)
-   include_directories(${SNDFILE_INCLUDE_DIR})
-   include_directories(${VORBIS_INCLUDE_DIR})
-   include_directories(${PORTAUDIO_INCLUDE_DIR})
-   include_directories(${OGG_INCLUDE_DIR})
-endif (MSVC)
-
-if (USE_SYSTEM_FREETYPE)
-      include_directories(${FREETYPE_INCLUDE_DIRS})
-else (USE_SYSTEM_FREETYPE)
-      include_directories(${PROJECT_SOURCE_DIR}/thirdparty/freetype/include)
-endif (USE_SYSTEM_FREETYPE)
-
-if (OMR)
- if (NOT USE_SYSTEM_POPPLER)
-   include_directories(${PROJECT_SOURCE_DIR}/thirdparty/poppler)
-   include_directories(${PROJECT_SOURCE_DIR}/thirdparty/poppler/poppler)
- else (NOT USE_SYSTEM_POPPLER)
-   include_directories(/usr/include/poppler)
-   include_directories(/usr/include/poppler/qt5)
- endif (NOT USE_SYSTEM_POPPLER)
-endif (OMR)
-
-add_subdirectory(mscore)
-
-if (BUILD_CRASH_REPORTER)
-      set (CRASHREPORTER_EXECUTABLE_NAME "${MSCORE_EXECUTABLE_NAME}-crash-reporter")
-      add_subdirectory(crashreporter)
-      add_dependencies(mscore mscore-crash-reporter)
-      set (CRASHREPORTER_EXECUTABLE "${CRASHREPORTER_EXECUTABLE_NAME}${CMAKE_EXECUTABLE_SUFFIX}")
-      set (MSCORE_EXECUTABLE "${MSCORE_EXECUTABLE_NAME}${CMAKE_EXECUTABLE_SUFFIX}")
-endif (BUILD_CRASH_REPORTER)
-
-##
-## produce config.h file
-##
-configure_file (
-      ${PROJECT_SOURCE_DIR}/build/config.h.in
-      ${PROJECT_BINARY_DIR}/config.h
-      )
 
 ##
 ##  Include packaging
