+++ mscore/musescore.cpp
@@ -16,6 +16,7 @@
 #include <QStyleFactory>
 #include "config.h"
 #include "musescore.h"
+#include "musescoredialogs.h"
 #include "scoreview.h"
 #include "libmscore/style.h"
 #include "libmscore/score.h"
@@ -56,6 +57,7 @@
 #include "keyedit.h"
 #include "harmonyedit.h"
 #include "navigator.h"
+#include "newwizard.h"
 #include "timeline.h"
 #include "importmidi/importmidi_panel.h"
 #include "importmidi/importmidi_instrument.h"
@@ -290,42 +292,6 @@ void MuseScore::cmdInsertMeasures()
                   insertMeasuresDialog->show();
                   }
             }
-      }
-
-//---------------------------------------------------------
-// InsertMeasuresDialog
-//---------------------------------------------------------
-
-InsertMeasuresDialog::InsertMeasuresDialog(QWidget* parent)
-   : QDialog(parent)
-      {
-      setObjectName("InsertMeasuresDialog");
-      setupUi(this);
-      setWindowFlags(this->windowFlags() & ~Qt::WindowContextHelpButtonHint);
-      setModal(true);
-      insmeasures->selectAll();
-      }
-
-//---------------------------------------------------------
-// Insert Measure -->   accept
-//---------------------------------------------------------
-
-void InsertMeasuresDialog::accept()
-      {
-      int n = insmeasures->value();
-      if (mscore->currentScore())
-            mscore->currentScoreView()->cmdInsertMeasures(n, ElementType::MEASURE);
-      done(1);
-      }
-
-//---------------------------------------------------------
-// InsertMeasuresDialog hideEvent
-//---------------------------------------------------------
-
-void InsertMeasuresDialog::hideEvent(QHideEvent* event)
-      {
-      MuseScore::saveGeometry(this);
-      QDialog::hideEvent(event);
       }
 
 //---------------------------------------------------------
@@ -2998,31 +2964,6 @@ void MuseScore::cmdAppendMeasures()
       }
 
 //---------------------------------------------------------
-//   MeasuresDialog
-//---------------------------------------------------------
-
-MeasuresDialog::MeasuresDialog(QWidget* parent)
-   : QDialog(parent)
-      {
-      setupUi(this);
-      setWindowFlags(this->windowFlags() & ~Qt::WindowContextHelpButtonHint);
-      setModal(true);
-      measures->selectAll();
-      }
-
-//---------------------------------------------------------
-//   accept
-//---------------------------------------------------------
-
-void MeasuresDialog::accept()
-      {
-      int n = measures->value();
-      if (mscore->currentScore())
-            mscore->currentScoreView()->cmdAppendMeasures(n, ElementType::MEASURE);
-      done(1);
-      }
-
-//---------------------------------------------------------
 //   rebuildAudioDrivers
 //---------------------------------------------------------
 
@@ -4740,84 +4681,6 @@ void MuseScore::about()
       AboutBoxDialog ab;
       ab.show();
       ab.exec();
-      }
-
-//---------------------------------------------------------
-//   AboutBoxDialog
-//---------------------------------------------------------
-
-AboutBoxDialog::AboutBoxDialog()
-      {
-      setupUi(this);
-      museLogo->setPixmap(QPixmap(preferences.isThemeDark() ?
-            ":/data/musescore-logo-transbg-m.png" : ":/data/musescore_logo_full.png"));
-
-      if (MuseScore::unstable())
-            versionLabel->setText(tr("Unstable Prerelease for Version: %1").arg(VERSION));
-      else {
-            auto msVersion = QString(VERSION) + QString(".") + QString(BUILD_NUMBER);// +QString(" Beta");
-            versionLabel->setText(tr("Version: %1").arg(msVersion));
-      }
-
-      revisionLabel->setText(tr("Revision: %1").arg(revision));
-      setWindowFlags(this->windowFlags() & ~Qt::WindowContextHelpButtonHint);
-
-      QString visitAndDonateString;
-#if !defined(FOR_WINSTORE)
-      visitAndDonateString = tr("Visit %1www.musescore.org%2 for new versions and more information.\nSupport MuseScore with your %3donation%4.")
-                                     .arg("<a href=\"https://www.musescore.org/\">")
-                                     .arg("</a>")
-                                     .arg("<a href=\"https://www.musescore.org/donate\">")
-                                     .arg("</a>");
-      visitAndDonateString += "\n\n";
-#endif
-      QString finalString = visitAndDonateString + tr("Copyright &copy; 1999-2019 Werner Schweer and Others.\nPublished under the GNU General Public License.");
-      finalString.replace("\n", "<br/>");
-      copyrightLabel->setText(QString("<span style=\"font-size:10pt;\">%1</span>").arg(finalString));
-      connect(copyRevisionButton, SIGNAL(clicked()), this, SLOT(copyRevisionToClipboard()));
-      copyRevisionButton->setIcon(*icons[int(Icons::copy_ICON)]);
-      }
-
-//---------------------------------------------------------
-//   copyRevisionToClipboard
-//---------------------------------------------------------
-
-void AboutBoxDialog::copyRevisionToClipboard()
-      {
-      QClipboard* cb = QApplication::clipboard();
-      QString sysinfo = "OS: ";
-      sysinfo += QSysInfo::prettyProductName();
-      sysinfo += ", Arch.: ";
-      sysinfo += QSysInfo::currentCpuArchitecture();
-      // endianness?
-      sysinfo += ", MuseScore version (";
-      sysinfo += QSysInfo::WordSize==32 ? "32" : "64";
-      auto msVersion = QString(VERSION) + QString(".") + QString(BUILD_NUMBER);
-      sysinfo += "-bit): " + msVersion + ", revision: ";
-      sysinfo += "github-musescore-musescore-";
-      sysinfo += revision;
-      cb->setText(sysinfo);
-      }
-
-//---------------------------------------------------------
-//   AboutBoxDialog
-//---------------------------------------------------------
-
-AboutMusicXMLBoxDialog::AboutMusicXMLBoxDialog()
-      {
-      setupUi(this);
-      setWindowFlags(this->windowFlags() & ~Qt::WindowContextHelpButtonHint);
-      label->setText(QString("<span style=\"font-size:10pt;\">%1<br/></span>")
-                     .arg(tr(   "MusicXML is an open file format for exchanging digital sheet music,\n"
-                                "supported by many applications.\n"
-                                "Copyright © 2004-2018 the Contributors to the MusicXML\n"
-                                "Specification, published by the W3C Music Notation Community\n"
-                                "Group under the W3C Community Contributor License Agreement\n"
-                                "(CLA):\n%1\n"
-                                "A human-readable summary is available:\n%2")
-                          .arg( "\n&nbsp;&nbsp;&nbsp;&nbsp;<a href=\"https://www.w3.org/community/about/agreements/cla/\">https://www.w3.org/community/about/agreements/cla/</a>\n",
-                                "\n&nbsp;&nbsp;&nbsp;&nbsp;<a href=\"https://www.w3.org/community/about/agreements/cla-deed/\">https://www.w3.org/community/about/agreements/cla-deed/</a>\n")
-                          .replace("\n","<br/>")));
       }
 
 //---------------------------------------------------------
@@ -7243,7 +7106,7 @@ void MuseScore::updateUiStyleAndTheme()
       Shortcut::refreshIcons();
       }
 
-MuseScoreApplication* MuseScoreApplication::initApplication(int argc, char** argv)
+MuseScoreApplication* MuseScoreApplication::initApplication(int& argc, char** argv)
       {
       QFile f(":/revision.h");
       f.open(QIODevice::ReadOnly);
@@ -7540,15 +7403,24 @@ MuseScoreApplication::CommandLineParseResult MuseScoreApplication::parseCommandL
       parseResult.argv = argv;
       return parseResult;
       }
-}
-
-using namespace Ms;
+} // namespace Ms
 
 //---------------------------------------------------------
-//   main
+//   initZitaResources
+///   must be placed outside of namespace
 //---------------------------------------------------------
 
-int main(int argc, char* av[])
+static void initZitaResources()
+      {
+      Q_INIT_RESOURCE(zita);
+      }
+
+namespace Ms {
+//---------------------------------------------------------
+//   runApplication
+//---------------------------------------------------------
+
+int runApplication(int& argc, char** av)
       {
 #ifndef NDEBUG
       qSetMessagePattern("%{file}:%{function}: %{message}");
@@ -7584,7 +7456,7 @@ int main(int argc, char* av[])
       QAccessible::installFactory(AccessibleSearchBox::SearchBoxFactory);
       QAccessible::installFactory(Awl::AccessibleAbstractSlider::AbstractSliderFactory);
 
-      Q_INIT_RESOURCE(zita);
+      initZitaResources();
 
 #ifndef Q_OS_MAC
       QSettings::setDefaultFormat(QSettings::IniFormat);
@@ -8038,3 +7910,4 @@ void MuseScore::scoreUnrolled(MasterScore * original)
       MasterScore * score = original->unrollRepeats();
       setCurrentScoreView(appendScore(score));
       }
+} // namespace Ms
